requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.5
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
fabric==3.2.2
fake==0.8
invoke==2.2.0
paramiko==3.4.0
psycopg2==2.9.9
pycparser==2.22
PyNaCl==1.5.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
sqlparse==0.5.0
wrapt==1.16.0
